+++ src/MobiFlightConnector/frontend/src/components/wizard/ConfigReferences/ConfigReferencePanel.tsx
