+++ src/MobiFlightConnector/frontend/src/components/wizard/ConfigWizard.tsx
@@ -1,7 +1,7 @@
 import ActionBindingPanel from "@/components/wizard/components/ActionBindingPanel"
-import ConfigReferencePanel from "@/components/wizard/components/ConfigReferencePanel"
+import ConfigReferencePanel from "@/components/wizard/ConfigReferences/ConfigReferencePanel"
 import ConfigTrigger from "@/components/wizard/components/ConfigTrigger"
-import PreconditionsPanel from "@/components/wizard/components/PreconditionsPanel"
+import PreconditionsPanel from "@/components/wizard/Precondition/PreconditionsPanel"
 import { IConfigItem } from "@/types"
 import { RefObject, useState } from "react"
 import { useNavigate, useSearchParams } from "react-router"
