+++ src/MobiFlightConnector/frontend/src/components/wizard/Precondition/PreconditionsPanel.tsx
